pyproject.toml
[project]
name = "human-mcp"
version = "0.1.0"
description = "A person-centric MCP server — your context, your rules. Thought experiment with synthetic data."
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
human-mcp = "src.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
